Add & use HOF helper, remove much whitespace
Many params used a similar pattern when setting state where the incoming
voltage parameter was divided by a coefficient (divisor). This
introduces a HOF to take the divisor as a parameter. It goes one (or
two) further and also assigns the param name to the corresponding index
in cmd_list, and the function for setting state (created by the HOF) to
the corresponding index in c2.

Additionally a bunch of whitespace has been deleted. It is definitely
less readable. Where comments could be inlined, they have been, but
there are a couple spots where this got ridiculous.

Anyway, this refactor brings the line count well below 400.